Require offsets in confirmation, allow disabling resending
In order to avoid situations where confirmations are accidentally done multiple times, the confirmation messages should contain the offsets that are to be confirmed. The easiest way to confirm messages is just to send back the offsets that were received earlier. The offsets can optionally be committed to Kafka using `commit` flag.

Resending can be disabled by setting resend timeout to zero.