Mostly consistency/style/spelling changes/fixes
The swapping of some companion object's apply method's calls is just for
readability.

The sbt-updates plugin can be used by running sbt dependencyUpdates and is quite
useful.

There is one TODO I've added that will either prove to be invalid in review or
needs fixing, so either the TODO can be removed or a fix added in a subsequent
commit.