KafkaConsumerActor wrapper (#35)
Wrapper type around Kafka consumer actor. Provides an API that isn't based on actor messages.

When creating a consumer actor using the apply method, an actor ref
factory should be passed in rather than the actor system. This allows
creating the consumer actor as a child actor instead of having to create
it at the top level of the actor system hierarchy.