Helpful typing around producer actor (#93)
* Let dead letter queue capture unhandled messages

* Type parameters in records toString

Show the type parameters of consumer and producer records in toString
methods to help debug which type parameters are carried in the types.

* Deprecate ProducerRecords.fromValues

There is no use-case for producing records with `Nothing` as the key or
value type. Users should instead use fromValuesWithKey to explicitly say
they don't want to use a key for the records.